Use saga to post the message
