+++ src/actions/index.js
@@ -1,12 +1,13 @@
 import { all } from 'redux-saga/effects';
 import watchRequestedMenu from './application';
 import { watchRequestedRolesWithCancel } from './roles';
-import { watchRequestedLoadSimpleConfigWithCancel } from './simple_config';
+import { watchRequestedLoadSimpleConfigWithCancel, watchPostedSimpleConfigWithCancel } from './simple_config';
 
 export default function* rootSaga() {
   yield all([
     watchRequestedRolesWithCancel(),
     watchRequestedLoadSimpleConfigWithCancel(),
+    watchPostedSimpleConfigWithCancel(),
     watchRequestedMenu(),
   ]);
 }
